Move some member function's to header 'identifier.hpp'
